Partial merge from 0.2.12 branch

+++ source/glest_game/type_instances/command.cpp
@@ -113,7 +113,7 @@ Command::Command(NetworkDataBuffer &buf) :
 void Command::save(XmlNode *node) const {
 	node->addChild("archetype", archetype);
 	node->addChild("type", type->getName());
-	node->addChild("flags", flags.flags);
+	node->addChild("flags", (int)flags.flags);
 	node->addChild("pos", pos);
 	node->addChild("pos2", pos2);
 	unitRef.save(node->addChild("unitRef"));
